+++ sdks/python/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.11.1] - 2025-06-05
 
+### Added
+
+- Adds `get_status` and `aio_get_status` methods to the `Runs` feature client, which return a workflow run's status by its ID.
+
 ### Changed
 
 - Fixes a couple of blocking calls buried in the admin client causing loop blockages on child spawning